Math/fixed: remove obsolete arithmetic wrappers

+++ src/Rough/RoughAltitude.hpp
@@ -72,7 +72,7 @@ public:
   }
 
   RoughAltitude &operator=(fixed other) {
-    value = (short)iround(other);
+    value = iround(other);
     return *this;
   }
